Refactor and restructure seismic profile tests
Moved seismic profile tests to a new `TestSeismicsProfiles` class for better organization and readability. Removed redundant imports, merged test logic, and improved file handling consistency. Marked certain tests to be run explicitly to streamline test execution flow.
