+++ tests/test_io/test_meshes/test_segy_reader.py
@@ -1,18 +1,14 @@
+import os
 import time
-
 from typing import List
 
+import matplotlib.pyplot as plt
+import numpy as np
 import pytest
-import os
 
 from subsurface import TriSurf, optional_requirements
 from subsurface.core.structs.base_structures import StructuredData, UnstructuredData
-import matplotlib.pyplot as plt
-import numpy as np
-
-from subsurface.modules.reader.profiles.profiles_core import create_vertical_mesh
 from subsurface.modules.reader.volume import segy_reader
-from subsurface.modules.reader.volume.segy_reader import apply_colormap_to_texture
 from subsurface.modules.visualization import to_pyvista_mesh, pv_plot, to_pyvista_grid
 from tests.conftest import RequirementsLevel
 
@@ -142,130 +138,9 @@ def test_plot_segy_as_struct_data_with_coords_dict(get_structured_data, get_imag
         time.sleep(2)
 
 
-def test_seismic_profile():
-    # filepath = os.getenv("PATH_TO_SEISMIC")
-    sd_array: StructuredData = segy_reader.read_in_segy(
-        filepath=(os.getenv("PATH_TO_SEISMIC_FINAL")),
-        ignore_geometry=True,
-        flip_y_axis=True
-    )
-
-    sd_array.active_data_array.plot()
-    plt.show(block=False)
 
 
-def test_seismic_profile_3D_from_segy():
-    filepath = os.getenv("PATH_TO_SEISMIC")
-    texture: StructuredData = segy_reader.read_in_segy(filepath, ignore_geometry=True)
-
-    # region coords
-    import pandas as pd
-    file_path = os.getenv("PATH_TO_SECTION")
-    df = pd.read_csv(
-        filepath_or_buffer=file_path,
-        skiprows=4,  # Skip the header lines above 'CDP'
-        delim_whitespace=True,  # Treat consecutive spaces as separators
-        names=["CDP", "X_COORD", "Y_COORD"]  # Assign column names
-    )
-
-    coords = df[["X_COORD", "Y_COORD"]].to_numpy()
-    # endregion
-
-    zmin = -450
-    zmax = 140
-    vertices, faces = create_vertical_mesh(coords, zmin, zmax)
-    geometry: UnstructuredData = UnstructuredData.from_array(vertices, faces)
-
-    texture = apply_colormap_to_texture(texture, cmap_name="bwr")
-    ts = TriSurf(
-        mesh=geometry,
-        texture=texture,
-        texture_origin=[coords[0][0], coords[0][1], zmin],
-        texture_point_u=[coords[-1][0], coords[-1][1], zmin],
-        texture_point_v=[coords[0][0], coords[0][1], zmax]
-    )
-
-    pv_plot(
-        meshes=[to_pyvista_mesh(ts)],
-        image_2d=False
-    )
-
-
-def test_seismic_profile_3D_from_interpreted_tiff():
-    filepath = os.getenv("PATH_TO_INTERPRETATION")
-
-    import tifffile as tiff  # Install with pip install tifffile
-
-    image = tiff.imread(filepath)
-
-    # Define the crop region: [y_start:y_end, x_start:x_end]
-    # Example: Crop a region starting at (100, 100) with a size of 200x300 (height x width)
-    y_start, y_end = 83, 730  # Vertical range
-    x_start, x_end = 60, 2080  # Horizontal range
-
-    # Perform the crop
-    cropped_image = image[y_start:y_end, x_start:x_end]
-    texture = StructuredData.from_numpy(cropped_image)
-
-    # region coords
-    import pandas as pd
-    file_path = os.getenv("PATH_TO_SECTION")
-    df = pd.read_csv(
-        filepath_or_buffer=file_path,
-        skiprows=4,  # Skip the header lines above 'CDP'
-        delim_whitespace=True,  # Treat consecutive spaces as separators
-        names=["CDP", "X_COORD", "Y_COORD"]  # Assign column names
-    )
-
-    coords = df[["X_COORD", "Y_COORD"]].to_numpy()
-    # endregion
-
-    zmin = -450
-    zmax = 140
-    vertices, faces = create_vertical_mesh(coords, zmin, zmax)
-    geometry: UnstructuredData = UnstructuredData.from_array(vertices, faces)
-
-    # texture = apply_colormap_to_texture(texture, cmap_name="bwr")
-    ts = TriSurf(
-        mesh=geometry,
-        texture=texture,
-        texture_origin=[coords[0][0], coords[0][1], zmin],
-        texture_point_u=[coords[-1][0], coords[-1][1], zmin],
-        texture_point_v=[coords[0][0], coords[0][1], zmax]
-    )
-
-    pv_plot(
-        meshes=[to_pyvista_mesh(ts)],
-        image_2d=False
-    )
-
-
-def test_interpreted_profile():
-    filepath = os.getenv("PATH_TO_INTERPRETATION")
-
-    import matplotlib.pyplot as plt
-    import tifffile as tiff  # Install with pip install tifffile
-
-    image = tiff.imread(filepath)
-
-    # Define the crop region: [y_start:y_end, x_start:x_end]
-    # Example: Crop a region starting at (100, 100) with a size of 200x300 (height x width)
-    y_start, y_end = 83, 730  # Vertical range
-    x_start, x_end = 60, 2080  # Horizontal range
-
-    # Perform the crop
-    cropped_image = image[y_start:y_end, x_start:x_end]
-
-    # Plot the cropped image
-    plt.figure(figsize=(8, 6))
-    plt.imshow(cropped_image, cmap='gray')  # Use 'gray' for grayscale images
-    plt.colorbar(label='Pixel Intensity')
-    plt.title("Cropped TIF Image")
-    plt.xlabel("Width (pixels)")
-    plt.ylabel("Height (pixels)")
-    plt.show()
-
-
+@pytest.mark.skip(reason="This test should only being run explicitly")
 def test_segy_3d_segy_seg() -> None:
     import xarray as xr
     from segysak.segy import segy_header_scan
@@ -297,9 +172,8 @@ def test_segy_3d_segy_seg() -> None:
     pass
 
 
-def test_segy_3d_segy_segsak_II() -> None:
+def test_segy_3d_segy__volume_segsak_II() -> None:
     import os
-    import matplotlib.pyplot as plt
     import xarray as xr
     from segysak.segy import segy_header_scan
